bilinear filter in Texture::getPixel

+++ src/graphics/texture.h
@@ -31,7 +31,7 @@ class LUMIX_ENGINE_API Texture : public Resource
 		void removeDataReference();
 		void onDataUpdated();
 		void save();
-		uint32_t getPixel(int x, int y);
+		uint32_t getPixel(float x, float y);
 
 	private:
 		void loaded(FS::IFile* file, bool success, FS::FileSystem& fs);
